separate out ajax call for gene autocompletion from generic autocompletion
