+++ client/src/components/elements/index.js
@@ -42,6 +42,7 @@ export { default as BiotypeSelect } from './BiotypeSelect';
 export { Page, PageLeft, PageMain, PageRight } from './Page';
 export { default as Timestamp } from './Timestamp';
 export { default as SimpleListPagination } from './SimpleListPagination';
+export { default as AutocompleteBase } from './AutocompleteBase';
 
 export const Dialog = withMobileDialog()(withStyles((theme) => ({
   paper: {
